move models, remove tests
