--- frontend/src/app/chat.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { ChatService } from './chat.service';
-
-describe('ChatService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [ChatService]
-    });
-  });
-
-  it('should be created', inject([ChatService], (service: ChatService) => {
-    expect(service).toBeTruthy();
-  }));
-});
