+++ frontend/src/app/chat.service.ts
@@ -4,7 +4,7 @@ import { HttpClient } from "@angular/common/http";
 import { environment } from "../environments/environment";
 import { catchError } from "rxjs/operators";
 import { HttpUtils } from "./util/http-utils";
-import { ChatMessage } from "./chat/chat-message";
+import { ChatMessage } from "./model/chat-message";
 
 @Injectable()
 export class ChatService {
