+++ frontend/src/app/in-memory-data.service.ts
@@ -1,5 +1,5 @@
 import { InMemoryDbService } from 'angular-in-memory-web-api';
-import { FEED_ITEMS } from "./feed/mock-feed-items";
+import { FEED_ITEMS } from "./model/mock-feed-items";
 
 export class InMemoryDataService implements InMemoryDbService {
