+++ frontend/src/app/model/chat-message.ts
