+++ frontend/src/app/model/fedd-item.ts
